Expose ProcessConfig and friends from the internal module
Fixes #62
